fix(UserList): Исправления
- Убрана конструкция forwardRef
- Изменено расположение пропсов
- Убрана бизнес логика из поля ввода
